+++ src/main/java/io/github/mfvanek/pg/index/maintenance/IndexesMaintenanceOnHost.java
@@ -12,13 +12,13 @@ package io.github.mfvanek.pg.index.maintenance;
 
 import io.github.mfvanek.pg.connection.HostAware;
 import io.github.mfvanek.pg.index.IndexesHealthAware;
-import io.github.mfvanek.pg.model.DuplicatedIndexes;
-import io.github.mfvanek.pg.model.ForeignKey;
-import io.github.mfvanek.pg.model.Index;
-import io.github.mfvanek.pg.model.IndexWithBloat;
-import io.github.mfvanek.pg.model.IndexWithNulls;
 import io.github.mfvanek.pg.model.PgContext;
-import io.github.mfvanek.pg.model.UnusedIndex;
+import io.github.mfvanek.pg.model.index.DuplicatedIndexes;
+import io.github.mfvanek.pg.model.index.ForeignKey;
+import io.github.mfvanek.pg.model.index.Index;
+import io.github.mfvanek.pg.model.index.IndexWithBloat;
+import io.github.mfvanek.pg.model.index.IndexWithNulls;
+import io.github.mfvanek.pg.model.index.UnusedIndex;
 
 import javax.annotation.Nonnull;
 import java.util.List;
