+++ src/main/java/io/github/mfvanek/pg/model/index/IndexWithNulls.java
@@ -8,7 +8,7 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.model;
+package io.github.mfvanek.pg.model.index;
 
 import io.github.mfvanek.pg.utils.Validators;
