+++ src/main/java/io/github/mfvanek/pg/model/table/TableNameAware.java
@@ -8,9 +8,10 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.model;
+package io.github.mfvanek.pg.model.table;
 
 import io.github.mfvanek.pg.common.health.logger.AbstractHealthLogger;
+import io.github.mfvanek.pg.model.index.IndexNameAware;
 
 import javax.annotation.Nonnull;
