+++ src/test/java/io/github/mfvanek/pg/common/health/DatabaseHealthImplTest.java
@@ -16,15 +16,15 @@ import io.github.mfvanek.pg.connection.HighAvailabilityPgConnectionImpl;
 import io.github.mfvanek.pg.connection.PgConnectionImpl;
 import io.github.mfvanek.pg.embedded.PostgresDbExtension;
 import io.github.mfvanek.pg.embedded.PostgresExtensionFactory;
-import io.github.mfvanek.pg.model.DuplicatedIndexes;
-import io.github.mfvanek.pg.model.ForeignKey;
-import io.github.mfvanek.pg.model.Index;
-import io.github.mfvanek.pg.model.IndexWithBloat;
-import io.github.mfvanek.pg.model.IndexWithNulls;
-import io.github.mfvanek.pg.model.Table;
-import io.github.mfvanek.pg.model.TableWithBloat;
-import io.github.mfvanek.pg.model.TableWithMissingIndex;
-import io.github.mfvanek.pg.model.UnusedIndex;
+import io.github.mfvanek.pg.model.index.DuplicatedIndexes;
+import io.github.mfvanek.pg.model.index.ForeignKey;
+import io.github.mfvanek.pg.model.index.Index;
+import io.github.mfvanek.pg.model.index.IndexWithBloat;
+import io.github.mfvanek.pg.model.index.IndexWithNulls;
+import io.github.mfvanek.pg.model.index.UnusedIndex;
+import io.github.mfvanek.pg.model.table.Table;
+import io.github.mfvanek.pg.model.table.TableWithBloat;
+import io.github.mfvanek.pg.model.table.TableWithMissingIndex;
 import io.github.mfvanek.pg.utils.DatabaseAwareTestBase;
 import io.github.mfvanek.pg.utils.DatabasePopulator;
 import org.junit.jupiter.api.Test;
