+++ src/test/java/io/github/mfvanek/pg/common/health/DatabaseHealthMultipleSchemasTest.java
@@ -10,18 +10,18 @@
 
 package io.github.mfvanek.pg.common.health;
 
-import io.github.mfvanek.pg.model.DuplicatedIndexes;
-import io.github.mfvanek.pg.model.ForeignKey;
-import io.github.mfvanek.pg.model.Index;
-import io.github.mfvanek.pg.model.IndexWithBloat;
-import io.github.mfvanek.pg.model.IndexWithNulls;
-import io.github.mfvanek.pg.model.IndexWithSize;
 import io.github.mfvanek.pg.model.PgContext;
-import io.github.mfvanek.pg.model.Table;
-import io.github.mfvanek.pg.model.TableNameAware;
-import io.github.mfvanek.pg.model.TableWithBloat;
-import io.github.mfvanek.pg.model.TableWithMissingIndex;
-import io.github.mfvanek.pg.model.UnusedIndex;
+import io.github.mfvanek.pg.model.index.DuplicatedIndexes;
+import io.github.mfvanek.pg.model.index.ForeignKey;
+import io.github.mfvanek.pg.model.index.Index;
+import io.github.mfvanek.pg.model.index.IndexWithBloat;
+import io.github.mfvanek.pg.model.index.IndexWithNulls;
+import io.github.mfvanek.pg.model.index.IndexWithSize;
+import io.github.mfvanek.pg.model.index.UnusedIndex;
+import io.github.mfvanek.pg.model.table.Table;
+import io.github.mfvanek.pg.model.table.TableNameAware;
+import io.github.mfvanek.pg.model.table.TableWithBloat;
+import io.github.mfvanek.pg.model.table.TableWithMissingIndex;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
