+++ src/test/java/io/github/mfvanek/pg/model/index/UnusedIndexTest.java
@@ -8,7 +8,7 @@
  * Licensed under the Apache License 2.0
  */
 
-package io.github.mfvanek.pg.model;
+package io.github.mfvanek.pg.model.index;
 
 import org.junit.jupiter.api.Test;
